requirements.txt
requests
opencv-python
flatbuffers
pillow
numpy